+++ src/main/java/cv/inps/rh/emprestimo/application/dto/HistoricoPagamentoDTO.java
@@ -9,6 +9,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -22,10 +23,10 @@ public class HistoricoPagamentoDTO  {
 
 
 
-  private Long valorTotalPago ;
+  private BigDecimal valorTotalPago ;
 
 
-  private Long saldoDivida ;
+  private BigDecimal saldoDivida ;
 
   @Valid
   private List<HistoricoPagamentoRowDTO> pagamentos = new ArrayList<>();
