+++ src/main/java/cv/inps/rh/emprestimo/application/dto/OutrosEmprestimosDTO.java
@@ -5,6 +5,7 @@ package cv.inps.rh.emprestimo.application.dto;
 
 import cv.igrp.framework.stereotype.IgrpDTO;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @IgrpDTO
@@ -16,7 +17,7 @@ public record OutrosEmprestimosDTO (
 
   LocalDate dataTermino,
 
-  Long valorEmprestimo,
+  BigDecimal valorEmprestimo,
 
-  Long valorPrestacaoMensal
+  BigDecimal valorPrestacaoMensal
 ){}
